+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixes:
 - Snap
     - The application crashes due to Snap API checking when snap is not installed
+    
+### UI:
+- Displaying a **verified** green icon next to a verified publisher's name
 
 ## [0.7.1] 2019-10-25
 ### Features